% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{field_weights}
\alias{field_weights}
\title{Construct field weights}
\usage{
field_weights(..., all = TRUE, use_default = TRUE)
}
\arguments{
\item{...}{Key-value pairs where the key is the name of the field and the
value is the weight.}

\item{all}{Whether to take all fields into account for weighting. Adds
\code{"_all"} to the query.}

\item{use_default}{Whether to use a sensible default when passing nothing
to \code{...}.}
}
\value{
A character vector of field weights.
}
\description{
Construct a vector that can be used to weight metadata fields of search hits.
To be used as input to the \code{fields} argument in
\code{\link{gesis_search}}.
}
\examples{
# use default weights
field_weights()

# set a stronger weight on the abstract
field_weights(abstract = 5)
}
