% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_get.R
\name{gesis_get}
\alias{gesis_get}
\title{Get a GESIS record}
\usage{
gesis_get(ids)
}
\arguments{
\item{ids}{Dataset IDs of the records.}
}
\value{
An object of class \code{gesis_record}.
}
\description{
Performs a simple lookup of one or multiple specific GESIS record ID and
retrieves their metadata record from the data archive.
}
\examples{
\donttest{# retrieve metadata on the ALLBUS microdata record
gesis_get("ZA5262")

# retrieve ALLBUS 2014, 2016, and 2018
gesis_get(c("ZA5240", "ZA5250", "ZA5270"))}
}
