% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_search.R
\name{gesis_records}
\alias{gesis_records}
\alias{format.gesis_record}
\alias{print.gesis_record}
\alias{format.gesis_records}
\alias{print.gesis_records}
\alias{as_gesis_records}
\title{Methods for \code{gesis_records} objects}
\usage{
\method{format}{gesis_record}(x, max_persons = 5, ...)

\method{print}{gesis_record}(x, max_persons = 5, ...)

\method{format}{gesis_records}(x, n = 3, max_persons = 5, ...)

\method{print}{gesis_records}(x, n = 5, max_persons = 5, ...)

as_gesis_records(x, ...)
}
\arguments{
\item{x}{An object of type \code{gesis_record} or \code{gesis_records}.
For \code{as_gesis_records}, a list-like consisting of objects of class
\code{gesis_record}.}

\item{max_persons}{Maximum number of authors to print. Defaults to 5.}

\item{...}{Arguments passed to the respective methods. Currently unused
in \code{as_gesis_records}.}

\item{n}{Maximum number of records to print. Defaults to 3.}
}
\value{
\code{print} methods return \code{x} invisibly. \code{format}
functions return a character string. \code{as_gesis_records} returns an
object of class \code{gesis_records}.
}
\description{
Methods and functions to print, format and cast to objects of class
\code{gesis_records}.
}
\examples{
\donttest{# wrap a list of records as gesis_records
records <- list(gesis_get("ZA7500"), gesis_get("ZA4789"))
as_gesis_records(records)}
}
