% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgesis-package.R
\docType{package}
\name{rgesis-package}
\alias{rgesis}
\alias{rgesis-package}
\title{rgesis: Interface to GESIS Search}
\description{
Provides programmatic access to the GESIS Search engine \url{https://search.gesis.org/} to retrieve metadata on research data, variables, literature, and tools. Authenticate and download datasets from the GESIS data archive with minimal dependencies.
}
\section{Global options}{

A number of global options can be set that change the behavior of package
functions. These include:

\describe{
\item{\code{rgesis_cache_disk}}{If \code{TRUE}, stores OAuth access tokens
in the \code{httr2} cache. This prevents repeated authentication using
credentials but stores access credentials on the disk. Defaults to
\code{FALSE}.}
\item{\code{rgesis_debug}}{Whether to echo GET requests before they are
sent. Defaults to FALSE.}
\item{\code{rgesis_download_purpose}}{Download purpose that must be
specified before downloading any kind of data from the GESIS catalogue.
This option is a way to specify a download purpose globally to prevent
having to pass it as an argument to each call of \code{\link{gesis_data}}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jslth/rgesis/}
  \item \url{https://jslth.github.io/rgesis/}
  \item Report bugs at \url{https://github.com/jslth/rgesis/issues}
}

}
\author{
\strong{Maintainer}: Jonas Lieth \email{jonas.lieth@gesis.org} (\href{https://orcid.org/0000-0002-3451-3176}{ORCID}) [copyright holder]

}
\keyword{internal}
