% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_files.R
\name{gesis_files}
\alias{gesis_files}
\alias{gesis_file_types}
\title{GESIS files}
\usage{
gesis_files(record, type = "dataset")

gesis_file_types(record)
}
\arguments{
\item{record}{Object of class \code{gesis_record} as returned by
\code{\link{gesis_search}} and \code{gesis_get} or dataset ID. If a dataset ID
is passed, the function performs a call to \code{\link{gesis_get}}.}

\item{type}{Type of data to download. The following values are supported:

\itemize{
\item{\code{dataset}: Survey dataset, usually in a Stata or SPSS data format}
\item{\code{questionnaire}: Survey questionnaire, usually in PDF format}
\item{\code{codebook}: Survey codebook, usually in PDF format}
\item{\code{otherdocs}: Other files like READMEs, method
reports, or variable lists, usually in PDF format}
\item{\code{uncategorized}: Other files that are not categorized, usually
used for external links to full texts (DOI, URN)}
}

Defaults to \code{"dataset"} because downloading PDF or HTML files rarely
makes sense in R. A list of available data types for a given record can be
retrieved using \code{\link{gesis_file_types}}.}
}
\value{
\code{gesis_files} returns a list of class \code{gesis_file}
containing information about the file name, format, size, download URL, and
whether the file needs a login to download.

\code{gesis_file_types} returns a character vector containing the file
types that are available. If \code{"none"}, files are available but not
categorized. If \code{NULL}, no files are available for download.
}
\description{
Explore the files associated with a GESIS Search record. The output of this
function may guide you in selecting a file to download using
\code{\link{gesis_data}}.

\code{gesis_file_types} gives an overview of the file categories, i.e.
datasets, questionnaires, codebooks or other documents, that are available
for a specific record.
}
\examples{
\donttest{# check what file types are available
gesis_file_types("ZA3753")

# show all dataset files for ALLBUS 1998
gesis_files("ZA3753")

# show all of its questionnaire files
gesis_files("ZA3753", type = "questionnaire")

# other record types can have files, too
type <- gesis_file_types("pretest-129")
gesis_files("pretest-129", type = "uncategorized")}
}
