% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_get.R
\name{gesis_get}
\alias{gesis_get}
\title{Get a GESIS record}
\usage{
gesis_get(id)
}
\arguments{
\item{id}{Dataset ID of the record.}
}
\value{
An object of class \code{gesis_record}.
}
\description{
Performs a simple lookup of a specific GESIS record ID and retrieves its
metadata record from the data archive.
}
\examples{
\donttest{# retrieve metadata on the ALLBUS microdata record
gesis_get("ZA5262")}
}
