% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_search.R
\name{gesis_search}
\alias{gesis_search}
\alias{as.data.frame.gesis_records}
\title{GESIS Search}
\usage{
gesis_search(
  query = NULL,
  type = NULL,
  default_operator = "AND",
  fields = field_weights(),
  sort = "relevance",
  publication_year = NULL,
  countries = NULL,
  topics = NULL,
  authors = NULL,
  sources = NULL,
  study_title = NULL,
  study_group = NULL,
  collection_year = NULL,
  thematic_collections = NULL,
  method_data_collection = NULL,
  method_sampling = NULL,
  method_time = NULL,
  method_unit = NULL,
  data_type = NULL,
  interview_language = NULL,
  pages = 1,
  tidy = FALSE
)

\method{as.data.frame}{gesis_records}(x, ...)
}
\arguments{
\item{query}{The search term to query. Terms can be seperated by operators
OR / AND. If no operator is specified, defaults to the operator specified
in \code{default_operator}. If \code{NULL}, searches for all records.
Queryless searches can still be filtered using the parameters below.}

\item{type}{Type of search results to retrieve. Can be one of
\code{"research_data"}, \code{"variables"}, \code{"instruments_tools"},
\code{"publication"}, \code{"gesis_bib"}, or \code{"gesis_web"} or
\code{"gesis_person"}. If \code{NULL}, includes all result types. Defaults to
\code{NULL}.}

\item{default_operator}{Character specifying the default operator to use
as seperator of search terms. Can be \code{"AND"} or \code{"OR"}. Defaults
to \code{"AND"}. Ignored if \code{query} is \code{NULL}.}

\item{fields}{Character vector of field weights as returned by
\code{\link{field_weights}}. Used to specify how much weight is given to
query matches in certain metadata fields. The higher the weight, the more
a field influences the match probability. Defaults to the default values
specified by \code{\link{field_weights}}. Ignored if \code{query} is
\code{NULL}.}

\item{sort}{Attribute to sort by. Must be one of \code{"relevance"},
\code{"newest_first"}, \code{"oldest_first"}, \code{"title_ascending"},
\code{"title_descending"}, \code{"variable_ascending"}, or
\code{"variable_descending"}. Defaults to \code{"relevance"}.}

\item{publication_year}{Numeric or character vector of length 2 specifying
the time range in which a record was published. The first value is the start
year and the second value is the end year. If \code{NULL}, returns records
from all years. Defaults to \code{NULL}.}

\item{countries}{Character vector specifying the country in which the record
was published. If \code{NULL}, returns records from all countries. Defaults
to \code{NULL}.}

\item{topics}{Character vector specifying the topics to filter by. Example
values include "Political behavior and attitudes" or "Migration". See
\url{https://search.gesis.org/} for a full list of possible topics.}

\item{authors}{Character vector specifying the authors to filter by.}

\item{sources}{Character vector specifying the source of the record.}

\item{study_title}{Character vector specifying the study title. Only
sensible for research data or variables.}

\item{study_group}{Character vector specifying the study group. Only
sensible for research data or variables.}

\item{collection_year}{Numeric or character vector of length 2 specifying
the time range in which a dataset was collected. The first value is the
start year and the second value is the end year. If \code{NULL}, returns
records from all years. Defaults to \code{NULL}. Only sensible for research
data or variables.}

\item{thematic_collections}{Character vector specifying the thematic
collection a dataset or variable belongs to. Example values include
"COVID-19" or "Environment and climate". See \url{https://search.gesis.org/}
for a full list of possible topics.}

\item{method_data_collection, method_sampling, method_time, method_unit}{Arguments
to filter by the methodology used in a study. It is possible to filter by
the mode of data collection (\code{method_data_collection}), sampling
procedure (\code{method_sampling}), temporal research design
(\code{method_unit}) and analysis unit (\code{method_unit}). Only sensible
for research data and variables.}

\item{data_type}{Character vector specifying the data type of a dataset or
variable. Example values include \code{"numeric"}, \code{"geospatial"},
and \code{"text"}. Only sensible for research data and variables.}

\item{interview_language}{Character vector specifying the interview
language of a variable. Usually only \code{"german"} and \code{"english"}
are available. Only sensible for variables.}

\item{pages}{Numeric vector specifying the page numbers of the search results.
By default, only the first page is returned. If \code{NULL}, iterates through
all pages.}

\item{tidy}{If \code{TRUE}, returns the records in a rectangular shape that
can be wrangled more easily. In this form, some more complex metadata fields
that cannot be converted to a rectangular shapre are discarded. If
\code{FALSE}, returns the records as a list-like \code{gesis_records} object.
This object is more comprehensive, but harder to work with. Defaults to
\code{FALSE}.}

\item{x}{Object of class \code{gesis_records}}

\item{...}{Arguments passed on to the respective S3 methods.}
}
\value{
If \code{tidy = FALSE}, returns an object of class
\code{gesis_records} consisting of multiple objects of class \code{gesis_record}.
Each \code{gesis_record} contains attributes \code{index}, \code{id}, and
\code{score} specifying the index, the record ID, and the match score.

If \code{tidy = TRUE}, returns a dataframe or tibble where each
row is a record and each column is a metadata field.

In any case, the query used to retrieve the data is stored in the attribute
\code{query}.
}
\description{
Query the GESIS search engine to retrieve metadata on research data,
variables, tools, publications, libraries, and websites.
}
\examples{
\donttest{# Make a queryless search across the entire database and return
# the first 10 results
gesis_search()

# Queries can be narrowed down by providing a query string
record <- gesis_search("climate change")

# gesis_search() searches for all types of records by default. If you only
# want to search for, say, datasets, you can set the `type` argument.
gesis_search(type = "research_data")

# By default, only 10 results (1 page) are returned. You can specify
# which pages should queried or if all should be returned
gesis_search("climate change", pages = 2:3) # pages 2 and 3
if (FALSE) gesis_search("climate change", pages = NULL) # all pages

# By default, results are wrapped in a complex list structure. This is
# good for representing detailed information, but not for analyzing data.
# Results can be tidied up to a dataframe using either `as_tibble()`
# or by setting the `tidy` argument to TRUE
tibble::as_tibble(record)
gesis_search("climate change", tidy = TRUE)

# Using the fields argument you can control how different text fields
# are weighted in matching the query string.
# Search for "climate change" only in texts, not in other fields
gesis_search("climate change", fields = field_weights(
  full_text = 10,
  abstract_en = 5,
  abstract = 3
))

# Search for articles in a specified journal
gesis_search("urban planning", fields = field_weights(coreJournalTitle = 10))

# Search strings are separated by an AND operator by default
# this default can be overwritten by setting default_operator = "OR"
# or by manually separating the query string
weight <- field_weights(title = 100)
and <- gesis_search(
  "climate change",
  sort = "title_ascending",
  fields = weight
)
or1 <- gesis_search(
  "climate OR change",
  sort = "title_ascending",
  fields = weight
)
or2 <- gesis_search(
  "climate change",
  sort = "title_ascending",
  default_operator = "OR",
  fields = weight
)
attr(and, "query") <- NULL
attr(or1, "query") <- NULL
attr(or2, "query") <- NULL
identical(and, or1)
identical(or1, or2)

# Similarly, you can construct complex query terms using operators
gesis_search("(climate or environmental) AND (change OR crisis OR action)")

# Another way to narrow down the search is to set filters. GESIS Search
# offers a number of filters, including topic, publication year or
# methodology. Sensible values can be checked on https://search.gesis.org/.
gesis_search(topics = "climate change", type = "publication")}
}
